Move CorrelatedSubQueries to dedicated file
Given that it is used both in LogicalPlanner and within the rule.
