+++ server/src/main/java/io/crate/planner/CorrelatedSubQueries.java
@@ -1,0 +1,56 @@
+/*
+ * Licensed to Crate.io GmbH ("Crate") under one or more contributor
+ * license agreements.  See the NOTICE file distributed with this work for
+ * additional information regarding copyright ownership.  Crate licenses
+ * this file to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.  You may
+ * obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * However, if you have executed another commercial license agreement
+ * with Crate these terms will supersede the license and you may use the
+ * software solely pursuant to the terms of the relevant commercial agreement.
+ */
+
+package io.crate.planner;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import org.jetbrains.annotations.Nullable;
+
+import io.crate.analyze.relations.QuerySplitter;
+import io.crate.expression.symbol.SelectSymbol;
+import io.crate.expression.symbol.Symbol;
+
+public record CorrelatedSubQueries(List<Symbol> correlatedSubQueries, List<Symbol> remainder) {
+
+    /**
+     * Splits the given Symbol tree into a list of symbols containing
+     * correlated sub-queries and a list of symbols without correlated
+     * subqueries (=remainder)
+     */
+    public static CorrelatedSubQueries of(@Nullable Symbol from) {
+        if (from == null) {
+            return new CorrelatedSubQueries(List.of(), List.of());
+        }
+        var values = QuerySplitter.split(from).values();
+        var remainder = new ArrayList<Symbol>(values.size());
+        var correlatedSubQueries = new ArrayList<Symbol>(values.size());
+        for (var symbol : values) {
+            if (symbol.any(s -> s instanceof SelectSymbol x && x.isCorrelated())) {
+                correlatedSubQueries.add(symbol);
+            } else {
+                remainder.add(symbol);
+            }
+        }
+        return new CorrelatedSubQueries(correlatedSubQueries, remainder);
+    }
+}
